Rename utilities/package_utils to utilities/package
