+++ imod/formats/prj/disv_conversion.py
@@ -17,7 +17,7 @@ import xugrid as xu
 
 import imod
 from imod.mf6.model import Modflow6Model
-from imod.mf6.utilities.package_utils import get_repeat_stress
+from imod.mf6.utilities.package import get_repeat_stress
 from imod.prepare.layer import get_upper_active_grid_cells
 from imod.typing import GridDataArray
 from imod.util.imports import MissingOptionalModule
